Restructure and fix old tests
