+++ .idea/digital-fuesim-manv.iml
@@ -2,9 +2,9 @@
 <module type="WEB_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/backend/test" isTestSource="true" />
       <sourceFolder url="file://$MODULE_DIR$/frontend/cypress" isTestSource="true" />
       <sourceFolder url="file://$MODULE_DIR$/shared/tests" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/backend/src/test" isTestSource="true" />
       <excludeFolder url="file://$MODULE_DIR$/.tmp" />
       <excludeFolder url="file://$MODULE_DIR$/temp" />
       <excludeFolder url="file://$MODULE_DIR$/tmp" />
