+++ backend/src/test/http-migration.spec.ts
@@ -1,5 +1,5 @@
 import fs from 'node:fs';
-import { UserReadableIdGenerator } from '../src/utils/user-readable-id-generator.js';
+import { UserReadableIdGenerator } from '../utils/user-readable-id-generator.js';
 import { createTestEnvironment } from './utils.js';
 
 const basePath = '../test-scenarios/migration-test-scenarios';
